content_creation_agency/requirements.txt
agency-swarm
openai>=1.0.0
python-dotenv
tavily-python
nltk>=3.8.1
requests>=2.31.0
google-api-python-client
google-auth-oauthlib
google-auth-httplib2
textblob
pandas
urllib3>=2.0.0
streamlit 